% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_cleaning.R
\name{as.numfactor}
\alias{as.numfactor}
\title{Convert factor values as exact numeric values}
\usage{
as.numfactor(x)
}
\arguments{
\item{x}{A \code{factor} variable (may contain NAs).}
}
\value{
A \code{numeric} vector.
}
\description{
Automatically convert the input \code{x (factor)} into a \code{numeric} vector while keeping the
exact value of \code{x}.
}
\details{
In \strong{R}, \emph{numeric factors} should be coded as \code{character} but it is rarely the
case because most functions actually prefer \code{factors}. It's one of the shortcomings of \strong{R}.
Consequently, there is no \strong{R} function that converts a \emph{numeric factor} (e.g. a binary
variable with only zeros and ones) into a \code{numeric} vector while keeping the correct value (if you use
\code{as.numeric(myfactor)}, you will get the underlying level codes, not the values as numbers), hence the
usefulness of \code{as.numfactor}.
}
\examples{
f <- as.factor(cars$speed)
as.numeric(f) # Converts the values into level ranks.
as.numfactor(f) # Works!
}
