% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_raw_data.R
\name{import_raw_data}
\alias{import_raw_data}
\title{Import the Raw Dataset}
\usage{
import_raw_data()
}
\value{
A dataframe
}
\description{
Import the raw CSV file of the 2020 tarping survey data.
}
\examples{
\dontrun{
mydata <- import_raw_data()
}
}
