% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_cleaning.R
\name{is_binary}
\alias{is_binary}
\title{Find Binary Variables}
\usage{
is_binary(x)
}
\arguments{
\item{x}{Any variable (vector, columns etc.) of any length.}
}
\value{
Logical (i.e. a TRUE or FALSE answer).
}
\description{
Automatically detect binary (boolean) variables, even if there is NAs in the data.
}
\examples{
xx <- c(0, 0, 1, 0, NA)
is_binary(xx) # TRUE
}
