% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jk.dusz.tarping-package.R
\docType{package}
\name{jk.dusz.tarping-package}
\alias{jk.dusz.tarping}
\alias{jk.dusz.tarping-package}
\title{jk.dusz.tarping: Prepare and Analyse the Tarping Survey Data}
\description{
This package project is meant to clean, prepare, and statistically analyse the data from 
    the 2020 Survey on tarping as a control method for Japanese knotweed s.l. (Reynoutria spp.), launched
    by Marie-Anne Dusz, André Evette, Fanny Dommanget and me.
}
\author{
\strong{Maintainer}: Francois-Marie Martin \email{francois-marie.martin@inrae.fr}

}
\keyword{internal}
