% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_cleaning.R
\name{lighten_my_data}
\alias{lighten_my_data}
\title{Lighter Version of the Data}
\usage{
lighten_my_data()
}
\value{
A tibble.
}
\description{
This function creates a slightly lighter version of the cleaned knotweed's
tarping survey dataset by removing the variables that will not be used in any of the future models
(such as purely descriptive variables).
}
\examples{
\dontrun{
my_light_dataset <- lighten_my_data()
}
}
