% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_preparation.R
\name{model_datasets}
\alias{model_datasets}
\title{Datasets for Models Building}
\usage{
model_datasets(
  response.var = c("efficiency", "edges", "overlaps", "tarpedarea")
)
}
\arguments{
\item{response.var}{A character string specifying which modelling dataset should be produced (either:
efficiency", "edges", "overlaps", or "tarpedarea"):
\itemize{
\item "efficiency" will produce the dataset for the 3 efficiency evaluation variables (namely \emph{
eff_eradication}, \emph{eff_expansion} and \emph{eff_vigour});
\item "edges" will produce the dataset having for response variable the tarping operations that observed
regrowth at the edge of the tarped area (at the end of the operation or during the latest visit to the
site);
\item "overlaps" produce the dataset having for response variable the tarping operations that observed
regrowth at strip overlaps;
\item "tarpedarea" will produce the dataset having for response variable the tarping operations that observed
regrowth at somewhere on the area covered by the fabric.
}}
}
\value{
A tibble.
}
\description{
The \code{model_datasets} function creates the respective reduced datasets that should be used
to model each response variables. For instance, if \code{response.var = "overlaps"}, the function will
produce a dataset containing \emph{reg_stripsoverlap} as the \strong{response} variable and a subset of
variables to be used as \strong{predictors/explanatory variables} (removing all the variables that
should not be used to model \emph{reg_stripsoverlap}'s variations).
}
\examples{
\dontrun{
eff_model <- model_datasets(response.var = "efficiency")
}
}
