% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_preparation.R
\name{uni.simudistrib}
\alias{uni.simudistrib}
\title{Random sample simulation}
\usage{
uni.simudistrib(simu.var, distribution)
}
\arguments{
\item{simu.var}{A \code{data.frame} or a \code{matrix}. For obvious layout and readability reasons, \code{simu.var}
should not include too many variables (\code{p} < 12 is advised) if the plot is to be printed in a page
or window of limited dimensions. For any use in HTML documents (e.g. with \code{RMarkdown}), the number of
input variables should not be a problem.}

\item{distribution}{Either \code{"normal"}, \code{"log-normal"} or \code{"poisson"} (case sensitive).}
}
\value{
A panel of \code{p}*5 plots, where \code{p} is the number of variables in \code{simu.var}.
}
\description{
The \code{uni.simudistrib} function \strong{automatically generates 5 Cleveland dotplots} of random
samples from different distributions (either \code{normal}, \code{log-normal}, or \code{poisson}) based on the
parameters of the variables in the input \code{data.frame} or \code{matrix} (see \emph{Details}).  \cr
This function is useful to see whether variables' extreme values are actual outliers or whether they
lie in a range of values possible for a random sample drawn from a \code{normal}, \code{log-normal}, or a \code{poisson}
distribution. \emph{In fine}, it may help determine if the original variable can be approximated by
these distribution with or without a transformation.
}
\details{
The \code{uni.simudistrib} function extracts some key parameters from the input variables (sample
size, mean and standard deviation) and generates random samples based on these parameters. For instance,
if \code{simu.var} contains \emph{i} variables \code{X1}, \code{X2}, ... \code{Xi} and if \code{distribution = "normal"}, the
function will return a panel of \emph{i}x5 plots:
\itemize{
\item The 1st row will contain five dotplots for five random samples with \emph{n} = \code{length(X1)} and drawn
from a Normal distribution with the same mean and standard deviation as \code{X1}.
\item The 2nd row will contain five dotplots for five random samples with \emph{n} = \code{length(X2)} and drawn
from a Normal distribution with the same mean and standard deviation as \code{X2}.
\item Etc.
}

\strong{Warning}: the function may fail for \code{log-normal} and \code{poisson} distributions if input
variables contain negative values (because these distributions are by definition positive). Additionally,
if \code{distribution = "poisson"}, the resulting plots will return \strong{integer} values as Poisson
is a discrete probability distribution.
}
\examples{
uni.simudistrib(simu.var = iris[,1:4], distribution = "normal")
}
