% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_raw_data.R
\name{import_raw_data}
\alias{import_raw_data}
\title{Import the Raw Dataset}
\usage{
import_raw_data()
}
\value{
A tibble (i.e. a kind of improved data.frame). For further information on tibbles, please refer to
the \link[tidyverse]{paper} or \link[readr]{readr} documentation.
}
\description{
Import the raw .csv file of the 2020 knotweeds' tarping survey data. \cr To avoid errors, please ensure
that there are no ";" in the cells of the CSV (e.g. if comments have been made in the table cells) and
that strings within the table are written in lower cases, without spaces or any special characters
(you can write in English for instance).
}
\examples{
\dontrun{
mydata <- import_raw_data()
}
}
